Revert "SERVER-43762 tighten the overload set for BSONObjBuilder::appendNumber"
This reverts commit 3307ba57eef2dfff9abe2413e39058ac9451ec04.

+++ src/mongo/db/pipeline/resume_token.cpp
@@ -105,7 +105,7 @@ ResumeToken::ResumeToken(const ResumeTokenData& data) {
     if (data.version >= 1) {
         builder.appendNumber("", data.tokenType);
     }
-    builder.appendNumber("", static_cast<long long>(data.txnOpIndex));
+    builder.appendNumber("", data.txnOpIndex);
     if (data.version >= 1) {
         builder.appendBool("", data.fromInvalidate);
     }
