+++ src/mongo/executor/connection_pool_stats.cpp
@@ -77,21 +77,20 @@ void ConnectionPoolStats::updateStatsForHost(std::string pool,
 }
 
 void ConnectionPoolStats::appendToBSON(mongo::BSONObjBuilder& result, bool forFTDC) {
-    result.appendNumber("totalInUse", static_cast<long long>(totalInUse));
-    result.appendNumber("totalAvailable", static_cast<long long>(totalAvailable));
-    result.appendNumber("totalCreated", static_cast<long long>(totalCreated));
-    result.appendNumber("totalRefreshing", static_cast<long long>(totalRefreshing));
+    result.appendNumber("totalInUse", totalInUse);
+    result.appendNumber("totalAvailable", totalAvailable);
+    result.appendNumber("totalCreated", totalCreated);
+    result.appendNumber("totalRefreshing", totalRefreshing);
 
     if (forFTDC) {
         BSONObjBuilder poolBuilder(result.subobjStart("connectionsInUsePerPool"));
         for (const auto& pool : statsByPool) {
             BSONObjBuilder poolInfo(poolBuilder.subobjStart(pool.first));
             auto& poolStats = pool.second;
-            poolInfo.appendNumber("poolInUse", static_cast<long long>(poolStats.inUse));
+            poolInfo.appendNumber("poolInUse", poolStats.inUse);
             for (const auto& host : poolStats.statsByHost) {
                 auto hostStats = host.second;
-                poolInfo.appendNumber(host.first.toString(),
-                                      static_cast<long long>(hostStats.inUse));
+                poolInfo.appendNumber(host.first.toString(), hostStats.inUse);
             }
         }
 
@@ -107,18 +106,18 @@ void ConnectionPoolStats::appendToBSON(mongo::BSONObjBuilder& result, bool forFT
         for (const auto& pool : statsByPool) {
             BSONObjBuilder poolInfo(poolBuilder.subobjStart(pool.first));
             auto& poolStats = pool.second;
-            poolInfo.appendNumber("poolInUse", static_cast<long long>(poolStats.inUse));
-            poolInfo.appendNumber("poolAvailable", static_cast<long long>(poolStats.available));
-            poolInfo.appendNumber("poolCreated", static_cast<long long>(poolStats.created));
-            poolInfo.appendNumber("poolRefreshing", static_cast<long long>(poolStats.refreshing));
+            poolInfo.appendNumber("poolInUse", poolStats.inUse);
+            poolInfo.appendNumber("poolAvailable", poolStats.available);
+            poolInfo.appendNumber("poolCreated", poolStats.created);
+            poolInfo.appendNumber("poolRefreshing", poolStats.refreshing);
 
             for (const auto& host : poolStats.statsByHost) {
                 BSONObjBuilder hostInfo(poolInfo.subobjStart(host.first.toString()));
                 auto& hostStats = host.second;
-                hostInfo.appendNumber("inUse", static_cast<long long>(hostStats.inUse));
-                hostInfo.appendNumber("available", static_cast<long long>(hostStats.available));
-                hostInfo.appendNumber("created", static_cast<long long>(hostStats.created));
-                hostInfo.appendNumber("refreshing", static_cast<long long>(hostStats.refreshing));
+                hostInfo.appendNumber("inUse", hostStats.inUse);
+                hostInfo.appendNumber("available", hostStats.available);
+                hostInfo.appendNumber("created", hostStats.created);
+                hostInfo.appendNumber("refreshing", hostStats.refreshing);
             }
         }
     }
@@ -127,10 +126,10 @@ void ConnectionPoolStats::appendToBSON(mongo::BSONObjBuilder& result, bool forFT
         for (auto&& host : statsByHost) {
             BSONObjBuilder hostInfo(hostBuilder.subobjStart(host.first.toString()));
             auto hostStats = host.second;
-            hostInfo.appendNumber("inUse", static_cast<long long>(hostStats.inUse));
-            hostInfo.appendNumber("available", static_cast<long long>(hostStats.available));
-            hostInfo.appendNumber("created", static_cast<long long>(hostStats.created));
-            hostInfo.appendNumber("refreshing", static_cast<long long>(hostStats.refreshing));
+            hostInfo.appendNumber("inUse", hostStats.inUse);
+            hostInfo.appendNumber("available", hostStats.available);
+            hostInfo.appendNumber("created", hostStats.created);
+            hostInfo.appendNumber("refreshing", hostStats.refreshing);
         }
     }
 }
